WIP on move year select dropdown to proposal create
