+++ app/views/proposal_types/years.turbo_stream.erb
@@ -1,0 +1,3 @@
+<%= turbo_stream.update @target do %>
+  <%= options_for_select @proposal_types_years %>
+<% end %>
