+++ package.json
@@ -8,6 +8,7 @@
     "@rails/actioncable": "^6.0.0",
     "@rails/actiontext": "^6.1.4",
     "@rails/activestorage": "^6.0.0",
+    "@rails/request.js": "^0.0.8",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "5.4.4",
     "@yaireo/tagify": "^4.8.1",
